+++ src/adapters/middlewares/validate-jwt.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import jwt from "jsonwebtoken";
-import { IPayloadJWT } from "../../../domain/entities/users";
+import { IPayloadJWT } from "../../domain/entities/users";
 
 export const validateJwt = (
   req: Request,
@@ -17,11 +17,11 @@ export const validateJwt = (
   }
 
   try {
-    const { uid, name }: IPayloadJWT = jwt.verify(
+    const { id, name }: IPayloadJWT = jwt.verify(
       token,
       process.env.SECRET_JWT_SEED as string
     ) as IPayloadJWT;
-    req.body.uid = uid;
+    req.body.id = id;
     req.body.name = name;
   } catch (error) {
     console.log(error);
